Chaning the sentence splitting annotator.

+++ build_script.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+#Edit this to have a parameter that is the pom.xml profile to use.
+
+if [ $# == 0 ]; then
+    echo $#
+    echo "Useage: build_script.sh PROFILE_NAME"
+    echo "  Please provide the profile to reference in the pom file."
+    exit 0;
+fi
+
 #Comment this thing. 
 [ -d "NLP_Platform_Installer" ] && rm -rf "NLP_Platform_Installer"
 [ ! -d "NLP_Platform_Installer" ] && git clone https://github.com/bholland/NLP_Platform_Installer.git
@@ -8,8 +17,11 @@
 
 ( [ ! -e "config.yaml" ] || [ ! -e "clean_xml.yaml" ] ) && python3 -m Main.main && echo "config.yaml and clean_xml.yaml files created. Edit these config files now under the current directory and rerun this file." && exit 0
 
+#clean and remove the target
 rm -rf target
-mvn -P project clean compile prepare-package package
+
+#Build the profile. Please note that $1 is the profile that the user passes.
+mvn -P $1 clean compile prepare-package package
 
 cp -rL models target/
 cp -r project_training_data target/
